various fixes
- exclude "data" folder in intellij/clion
- make errors in minecraft launcher launcher send desktop notifications
- fix release_build_fast_install_and_run.sh script not installing after recent changes
- fix clippy warnings in rng_simulator.rs
- run shellcheck & tests in run.sh
- use Lazy<Arc<AtomicBool>> instead of just AtomicBool
- remove unnecessary async/await and tokio::fs usage, hopefully fixes deadlock issue

I think the deadlock happened cause we got a new modify event while the file for previous event was being saved and that event we got tried to save too and so the .await's deadlocked somehow. Not sure though, maybe it was something else, I thought it was the Mutex#lock in last commits but it happened after removing mutex too. If this still doesn't fix it then I think it's cause of AtomicBool.

- fix tests after recent changes to rng_simulator.rs