mll: launch with higher priority
MLL (Minecraft Launcher Launcher): Launch with a default nice priority of -6, which is higher than all other processes except kernel and audio processing (pipewire and pulse), at least on my system. This makes it so that if browser or discord, or any other program is open while the game is running, game is prioritized.