use a backup-restore solution
.. for the issue of launcher profiles resetting. There is no better way.

First we were killing with Rust's kill() method, so we thought if we switched over to SIGTERM instead of SIGKILL, the launcher would save the files/finish the IO before exiting, but turns out the launcher doesn't handle any signals at all, it only handles XDestroyEvent, i.e closing via clicking the X button from launcher in the desktop.

This can be simulated programmatically via wmctrl command/crate, HOWEVER, this only works when the game is not open. Otherwise it will just make the launcher keep working on background but close the window.

Therefore the only solution is to backup the file, use SIGTERM (doesn't matter but more idiomatic instead of SIGKILL), then restore the file.

At least until Mojang adds a new setting to completely close the launcher after opening the game or handles the SIGKILL signal.